Revert "refactor: refactor transaction functions"
