+++ MoneyEz.API/Controllers/TransactionsController.cs
@@ -22,18 +22,11 @@ namespace MoneyEz.API.Controllers
     public class TransactionsController : BaseController
     {
         private readonly ITransactionService _transactionService;
-        private readonly IGroupTransactionService _groupTransactionService;
-        private readonly IExternalTransactionService _externalTransactionService;
         private readonly IClaimsService _claimsService;
 
-        public TransactionsController(ITransactionService transactionService, 
-            IGroupTransactionService groupTransactionService,
-            IExternalTransactionService externalTransactionService,
-            IClaimsService claimsService)
+        public TransactionsController(ITransactionService transactionService, IClaimsService claimsService)
         {
             _transactionService = transactionService;
-            _groupTransactionService = groupTransactionService;
-            _externalTransactionService = externalTransactionService;
             _claimsService = claimsService;
         }
 
@@ -73,12 +66,6 @@ namespace MoneyEz.API.Controllers
         {
             return ValidateAndExecute(() => _transactionService.DeleteTransactionAsync(transactionId));
         }
-
-        [HttpPut("categorize")]
-        public Task<IActionResult> CategorizeTransaction(CategorizeTransactionModel model)
-        {
-            return ValidateAndExecute(() => _transactionService.CategorizeTransactionAsync(model));
-        }
         #endregion single user
 
         #region admin   
@@ -97,38 +84,38 @@ namespace MoneyEz.API.Controllers
         public Task<IActionResult> GetAllTransactionsForGroupByGroupID([FromQuery] Guid groupId, 
             [FromQuery] PaginationParameter paginationParameter, [FromQuery] TransactionFilter filter)
         {
-            return ValidateAndExecute(() => _groupTransactionService.GetTransactionByGroupIdAsync(groupId, paginationParameter, filter));
+            return ValidateAndExecute(() => _transactionService.GetTransactionByGroupIdAsync(groupId, paginationParameter, filter));
         }
 
         [HttpPost("group/create")]
         public Task<IActionResult> CreateGroupTransaction([FromBody] CreateGroupTransactionModel model)
         {
             var currentEmail = _claimsService.GetCurrentUserEmail;
-            return ValidateAndExecute(() => _groupTransactionService.CreateGroupTransactionAsync(model, currentEmail));
+            return ValidateAndExecute(() => _transactionService.CreateGroupTransactionAsync(model, currentEmail));
         }
 
         [HttpGet("group/transaction/{transactionId}")]
         public Task<IActionResult> GetGroupTransactionDetails(Guid transactionId)
         {
-            return ValidateAndExecute(() => _groupTransactionService.GetGroupTransactionDetailsAsync(transactionId));
+            return ValidateAndExecute(() => _transactionService.GetGroupTransactionDetailsAsync(transactionId));
         }
 
         [HttpPut("group/update")]
         public Task<IActionResult> UpdateGroupTransaction([FromBody] UpdateGroupTransactionModel model)
         {
-            return ValidateAndExecute(() => _groupTransactionService.UpdateGroupTransactionAsync(model));
+            return ValidateAndExecute(() => _transactionService.UpdateGroupTransactionAsync(model));
         }
 
         [HttpDelete("group/delete/{transactionId}")]
         public Task<IActionResult> DeleteGroupTransaction(Guid transactionId)
         {
-            return ValidateAndExecute(() => _groupTransactionService.DeleteGroupTransactionAsync(transactionId));
+            return ValidateAndExecute(() => _transactionService.DeleteGroupTransactionAsync(transactionId));
         }
 
         [HttpPost("group/response")]
         public Task<IActionResult> ResponseGroupTransaction(ResponseGroupTransactionModel model)
         {
-            return ValidateAndExecute(() => _groupTransactionService.ResponseGroupTransactionAsync(model));
+            return ValidateAndExecute(() => _transactionService.ResponseGroupTransactionAsync(model));
         }
 
         #region vote
@@ -136,38 +123,44 @@ namespace MoneyEz.API.Controllers
         [HttpPost("group/vote/create")]
         public Task<IActionResult> CreateGroupTransactionVote([FromBody] CreateGroupTransactionVoteModel model)
         {
-            return ValidateAndExecute(() => _groupTransactionService.CreateGroupTransactionVoteAsync(model));
+            return ValidateAndExecute(() => _transactionService.CreateGroupTransactionVoteAsync(model));
         }
 
         [HttpPut("group/vote/update")]
         public Task<IActionResult> UpdateGroupTransactionVote([FromBody] UpdateGroupTransactionVoteModel model)
         {
-            return ValidateAndExecute(() => _groupTransactionService.UpdateGroupTransactionVoteAsync(model));
+            return ValidateAndExecute(() => _transactionService.UpdateGroupTransactionVoteAsync(model));
         }
 
         [HttpDelete("group/vote/delete/{voteId}")]
         public Task<IActionResult> DeleteGroupTransactionVote(Guid voteId)
         {
-            return ValidateAndExecute(() => _groupTransactionService.DeleteGroupTransactionVoteAsync(voteId));
+            return ValidateAndExecute(() => _transactionService.DeleteGroupTransactionVoteAsync(voteId));
         }
 
         #endregion vote
 
         #endregion group
 
-        #region external transaction
+        #region duc
         [HttpPost("webhook")]
         public Task<IActionResult> UpdateTransactionWebhook(WebhookPayload webhookPayload)
         {
-            return ValidateAndExecute(() => _externalTransactionService.UpdateTransactionWebhook(webhookPayload));
+            return ValidateAndExecute(() => _transactionService.UpdateTransactionWebhook(webhookPayload));
         }
 
         [HttpPost("python-service")]
         public Task<IActionResult> CreateTransactionPythonService(CreateTransactionPythonModel model)
         {
-            return ValidateAndExecute(() => _externalTransactionService.CreateTransactionPythonService(model));
+            return ValidateAndExecute(() => _transactionService.CreateTransactionPythonService(model));
         }
-        #endregion
+
+        [HttpPut("categorize")]
+        public Task<IActionResult> CategorizeTransaction(CategorizeTransactionModel model)
+        {
+            return ValidateAndExecute(() => _transactionService.CategorizeTransactionAsync(model));
+        }
+        #endregion duc
 
         #region report
