+++ MoneyEz.Services/Services/Interfaces/ITransactionService.cs
@@ -23,7 +23,31 @@ namespace MoneyEz.Services.Services.Interfaces
         #endregion personal
 
         Task<BaseResultModel> GetAllTransactionsForAdminAsync(PaginationParameter paginationParameter, TransactionFilter transactionFilter);
+
+        #region group
+        Task<BaseResultModel> GetTransactionByGroupIdAsync(Guid groupId,
+            PaginationParameter paginationParameter, TransactionFilter transactionFilter);
         Task<BaseResultModel> CategorizeTransactionAsync(CategorizeTransactionModel model);
+        Task<BaseResultModel> CreateGroupTransactionAsync(CreateGroupTransactionModel model, string currentEmail);
+        Task<BaseResultModel> GetGroupTransactionDetailsAsync(Guid transactionId);
+        Task<BaseResultModel> UpdateGroupTransactionAsync(UpdateGroupTransactionModel model);
+        Task<BaseResultModel> DeleteGroupTransactionAsync(Guid transactionId);
+
+        Task<BaseResultModel> ResponseGroupTransactionAsync(ResponseGroupTransactionModel model);
+        Task<BaseResultModel> RejectGroupTransactionAsync(Guid transactionId);
+        Task<BaseResultModel> CreateGroupTransactionVoteAsync(CreateGroupTransactionVoteModel model);
+        Task<BaseResultModel> UpdateGroupTransactionVoteAsync(UpdateGroupTransactionVoteModel model);
+        Task<BaseResultModel> DeleteGroupTransactionVoteAsync(Guid voteId);
+
+        #endregion group
+
+        #region python webhook
+        Task<BaseResultModel> UpdateTransactionWebhook(WebhookPayload webhookPayload);
+        Task<BaseResultModel> CreateTransactionPythonService(CreateTransactionPythonModel createTransactionPythonModel);
+        Task<BaseResultModel> CreateTransactionPythonServiceV2(CreateTransactionPythonModelV2 createTransactionPythonModel);
+        Task<BaseResultModel> GetTransactionHistorySendToPythons(Guid userId, TransactionFilter transactionFilter);
+
+        #endregion python webhook
 
         #region report
         Task<BaseResultModel> GetYearReportAsync(int year, string type);
